requirements.txt
bert-score
datasets
newspaper3k
nltk
numpy
opencv-python
pandas
Pillow
py7zr
pytrec-eval
ray
regex
requests
Scrapy
sentence-transformers
spacy
spacy-legacy
tensorboard
tensorboard-data-server
tensorboard-plugin-wit
tensorboardX
tokenizers
torch
torchaudio
torchvision
tqdm
transformers
typer
ujson
urllib3
wandb
hydra-core
python-dotenv
openai
accelerate
trl
bitsandbytes
peft
sentencepiece
flash_attn
